feat: extablished connection to db
