+++ README.md
@@ -1,2 +1,2 @@
-# boo
-kinda test .....
+# Boo
+kinda test - [Boo World](https://boo.world).
