+++ in-memory-error.txt
@@ -1,0 +1,12 @@
+Server started. Listening on 3000
+Starting the MongoMemoryServer Instance failed, enable debug log for more information. Error:
+ StdoutInstanceError: Instance failed to start because a library is missing or cannot be opened: "libcrypto.so.1.1"
+    at MongoInstance.checkErrorInLine (/home/pacifiquem/Documents/projects/self/boo/node_modules/mongodb-memory-server-core/lib/util/MongoInstance.js:382:62)
+    at MongoInstance.stderrHandler (/home/pacifiquem/Documents/projects/self/boo/node_modules/mongodb-memory-server-core/lib/util/MongoInstance.js:304:14)
+    at Socket.emit (node:events:513:28)
+    at addChunk (node:internal/streams/readable:315:12)
+    at readableAddChunk (node:internal/streams/readable:289:9)
+    at Socket.Readable.push (node:internal/streams/readable:228:10)
+    at Pipe.onStreamRead (node:internal/stream_base_commons:190:23)
+Connection to DataBase failed:  Instance failed to start because a library is missing or cannot be opened: "libcrypto.so.1.1"
+[nodemon] app crashed - waiting for file changes before starting...
