+++ package.json
@@ -4,7 +4,7 @@
   "description": "",
   "main": "app.js",
   "scripts": {
-    "dev":"nodemon app.js",
+    "dev": "nodemon app.js",
     "start": "node app.js"
   },
   "keywords": [],
@@ -14,6 +14,8 @@
     "dotenv": "^16.3.1",
     "ejs": "^3.1.6",
     "express": "^4.17.1",
+    "mongodb-memory-server": "^8.13.0",
+    "mongoose": "^7.3.4",
     "nodemon": "^3.0.1"
   }
 }
