+++ utils/profileData.utils.js
@@ -1,0 +1,15 @@
+module.exports = profiles = [
+    {
+      "id": 1,
+      "name": "A Martinez",
+      "description": "Adolph Larrue Martinez III.",
+      "mbti": "ISFJ",
+      "enneagram": "9w3",
+      "variant": "sp/so",
+      "tritype": 725,
+      "socionics": "SEE",
+      "sloan": "RCOEN",
+      "psyche": "FEVL",
+      "image": "https://soulverse.boo.world/images/1.png",
+    }
+];
